# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "Gomoku API", "description": "API para o jogo Gomoku - Projeto Web UFSC", "version": "1.0.0"}, "paths": {"/api/auth/register": {"post": {"summary": "Register", "operationId": "register_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/login": {"post": {"summary": "Login", "operationId": "login_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/me": {"get": {"summary": "Get Current User Info", "operationId": "get_current_user_info_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/users/": {"get": {"summary": "Get Users", "description": "Get list of all users", "operationId": "get_users__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/User"}, "type": "array", "title": "Response Get Users  Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/users/{user_id}": {"get": {"summary": "Get User", "description": "Get user by ID", "operationId": "get_user__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/me": {"put": {"summary": "Update Current User", "description": "Update current user profile", "operationId": "update_current_user_me_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/online/count": {"get": {"summary": "Get Online Users Count", "description": "Get count of online users", "operationId": "get_online_users_count_online_count_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/games/create": {"post": {"summary": "Create Game", "description": "Create a new game with specified mode and difficulty", "operationId": "create_game_create_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateGameRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GameResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/games/": {"get": {"summary": "Get Games", "description": "Get all games for the current user", "operationId": "get_games__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/games/{game_id}": {"get": {"summary": "Get Game", "description": "Get a specific game by ID", "operationId": "get_game__game_id__get", "parameters": [{"name": "game_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Game Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/games/{game_id}/move": {"post": {"summary": "Make Move", "description": "Make a move in a game", "operationId": "make_move__game_id__move_post", "parameters": [{"name": "game_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Game Id"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MoveRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/games/{game_id}/leave": {"post": {"summary": "Leave Game", "description": "Handle a user leaving a game.", "operationId": "leave_game__game_id__leave_post", "parameters": [{"name": "game_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Game Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ws/game/{game_id}": {"get": {"summary": "Websocket Game Endpoint", "description": "WebSocket endpoint for real-time game communication", "operationId": "websocket_game_endpoint_ws_game__game_id__get", "parameters": [{"name": "game_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Game Id"}}, {"name": "token", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/lobby/games": {"get": {"summary": "Get Active Games", "description": "Get list of active games.", "operationId": "get_active_games_games_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/lobby/players": {"get": {"summary": "Get Online Players", "description": "Get list of online players.", "operationId": "get_online_players_players_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/health": {"get": {"summary": "Health Check", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"CreateGameRequest": {"properties": {"mode": {"type": "string", "title": "Mode"}, "difficulty": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "medium", "title": "Difficulty"}}, "type": "object", "required": ["mode"], "title": "CreateGameRequest"}, "GameResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "mode": {"type": "string", "title": "Mode"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "mode", "status", "created_at"], "title": "GameResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Location": {"properties": {"city": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "", "title": "City"}, "state": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "", "title": "State"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "", "title": "Country"}}, "type": "object", "title": "Location"}, "LoginRequest": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "MoveRequest": {"properties": {"row": {"type": "integer", "title": "Row"}, "col": {"type": "integer", "title": "Col"}}, "type": "object", "required": ["row", "col"], "title": "MoveRequest"}, "RegisterRequest": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}, "name": {"type": "string", "title": "Name"}, "age": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Age"}, "city": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "", "title": "City"}, "state": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "", "title": "State"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "default": "", "title": "Country"}}, "type": "object", "required": ["username", "email", "password", "name"], "title": "RegisterRequest"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}, "user": {"$ref": "#/components/schemas/UserPublic"}}, "type": "object", "required": ["access_token", "token_type", "user"], "title": "Token"}, "User": {"properties": {"id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Id"}, "username": {"type": "string", "title": "Username"}, "email": {"type": "string", "format": "email", "title": "Email"}, "is_active": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "default": true, "title": "Is Active"}, "created_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Created At"}, "last_login": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Last Login"}, "profile": {"$ref": "#/components/schemas/UserProfile"}, "stats": {"$ref": "#/components/schemas/UserStats"}}, "type": "object", "required": ["username", "email", "profile", "stats"], "title": "User"}, "UserProfile": {"properties": {"name": {"type": "string", "title": "Name"}, "age": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Age"}, "location": {"anyOf": [{"$ref": "#/components/schemas/Location"}, {"type": "null"}], "title": "Location"}, "avatar_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Avatar Url"}}, "type": "object", "required": ["name"], "title": "UserProfile"}, "UserPublic": {"properties": {"id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Id"}, "username": {"type": "string", "title": "Username"}, "email": {"type": "string", "format": "email", "title": "Email"}, "is_active": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "default": true, "title": "Is Active"}, "created_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Created At"}, "last_login": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Last Login"}, "profile": {"$ref": "#/components/schemas/UserProfile"}, "stats": {"$ref": "#/components/schemas/UserStats"}}, "type": "object", "required": ["username", "email", "profile", "stats"], "title": "UserPublic"}, "UserStats": {"properties": {"games_played": {"anyOf": [{"type": "integer"}, {"type": "null"}], "default": 0, "title": "Games Played"}, "games_won": {"anyOf": [{"type": "integer"}, {"type": "null"}], "default": 0, "title": "Games Won"}, "games_lost": {"anyOf": [{"type": "integer"}, {"type": "null"}], "default": 0, "title": "Games Lost"}, "current_score": {"anyOf": [{"type": "integer"}, {"type": "null"}], "default": 1200, "title": "Current Score"}}, "type": "object", "title": "UserStats"}, "UserUpdate": {"properties": {"username": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Username"}, "email": {"anyOf": [{"type": "string", "format": "email"}, {"type": "null"}], "title": "Email"}, "profile": {"anyOf": [{"$ref": "#/components/schemas/UserProfile"}, {"type": "null"}], "title": "Profile"}}, "type": "object", "title": "UserUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}